% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_network.R
\name{compute_network}
\alias{compute_network}
\title{compute_network}
\usage{
compute_network(x, mode = "undirected")
}
\arguments{
\item{x}{An object of class \code{swaglm}.}

\item{mode}{Character string specifying the mode of the network. Default is "undirected".}
}
\value{
test
}
\description{
Compute a network representation of the selected models from a \code{swaglm} object
}
