% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_toeplitz_matrix_from_vector_cpp}
\alias{fast_toeplitz_matrix_from_vector_cpp}
\title{Generate a Toeplitz Matrix from a Vector}
\usage{
fast_toeplitz_matrix_from_vector_cpp(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
A numeric matrix that is the Toeplitz matrix generated from \code{v}.
}
\description{
This function generates a Toeplitz matrix from a given vector using the Armadillo library.
}
