% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{postprocessing_swaglm}
\alias{postprocessing_swaglm}
\title{postprocessing_swaglm}
\usage{
postprocessing_swaglm(x)
}
\arguments{
\item{x}{An object of class \code{swaglm}.}
}
\value{
A list with three elements:
\describe{
\item{mat_selected_model}{A matrix where each row represents a selected model and the columns indicate the indices of the variables included in that model.}
\item{vec_aic_selected_model}{A numeric vector containing the AIC values corresponding to each selected model in \code{mat_selected_model}.}
\item{estimated_beta_per_variable}{A named list where each element corresponds to a variable (named \code{V<index>}) and contains a numeric vector of estimated beta coefficients for that variable across all selected models in which it appears.}
}
}
\description{
postprocessing_swaglm
}
\examples{
set.seed(12345)
n <- 2000
p <- 100
# create design matrix and vector of coefficients
Sigma <- diag(rep(1/p, p))
X <- MASS::mvrnorm(n = n, mu = rep(0, p), Sigma = Sigma)
beta = c(-15,-10,5,10,15, rep(0,p-5))

# --------------------- generate from logistic regression with an intercept of one
z <- 1 + X\%*\%beta
pr <- 1/(1 + exp(-z))
y <- as.factor(rbinom(n, 1, pr))
y = as.numeric(y)-1

# define swag parameters
quantile_alpha = .15
p_max = 20
swaglm_obj = swaglm::swaglm(X=X, y = y, p_max = p_max, family = stats::binomial(),
                          alpha = quantile_alpha, verbose = TRUE, seed = 123)
swaglm_obj
swaglm_postprocess = postprocessing_swaglm(swaglm_obj)

# plot distribution of estimated beta with respect to true value
for(i in seq_along(swaglm_postprocess$estimated_beta_per_variable)){
  # get variable name
  var_name_i = names(swaglm_postprocess$estimated_beta_per_variable)[i]
  var_index_i =  as.numeric(substr(var_name_i, 2, nchar(var_name_i)))
  boxplot(swaglm_postprocess$estimated_beta_per_variable[[i]], 
          main=paste0(var_name_i, " ", "true value=", beta[i]))
}
}
