% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaglm_test.R
\name{print.swaglm_test}
\alias{print.swaglm_test}
\title{print.swaglm_test}
\usage{
\method{print}{swaglm_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{swaglm_test}.}

\item{...}{Additional arguments}
}
\description{
Print a \code{swaglm_test} object
}
