% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_estimation_model_one_dimension_cpp}
\alias{run_estimation_model_one_dimension_cpp}
\title{Run Estimation Model in One Dimension}
\usage{
run_estimation_model_one_dimension_cpp(X, y, family = NULL, method = 0L)
}
\description{
This function runs an estimation model in one dimension using the fastglm function.
}
