% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swag_test.R
\name{swag_test}
\alias{swag_test}
\title{significance_test}
\usage{
swag_test(swag_obj, significance_level = 0.05, B = 50, verbose = FALSE)
}
\arguments{
\item{swag_obj}{An object of class \code{swaglm}.}

\item{significance_level}{A \code{double} between 0 and 1 indicating the significance level. Default is 0.05.}

\item{B}{a \code{integer} specifying the number of swaglm procedure to perform under the null.}

\item{verbose}{A boolean xxx}
}
\description{
Compute significance of identified set of variables
}
\examples{
n <- 2000
p <- 50

# create design matrix and vector of coefficients
Sigma <- diag(rep(1/p, p))
X <- MASS::mvrnorm(n = n, mu = rep(0, p), Sigma = Sigma)
beta = c(-15,-10,5,10,15, rep(0,p-5))
z <- 1 + X\%*\%beta
pr <- 1/(1 + exp(-z))
y <- as.factor(rbinom(n, 1, pr))
y = as.numeric(y)-1
quantile_alpha = .15
p_max = 20
swag_obj = swaglm::swaglm(X=X, y = y, p_max = p_max, family = stats::binomial(),
                          alpha = quantile_alpha, verbose = TRUE, seed = 123)
swag_test(swag_obj, significance_level = .05, B = 50, verbose = TRUE)

}
