% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_network.R
\name{compute_network}
\alias{compute_network}
\title{compute_network}
\usage{
compute_network(x, mode = "undirected")
}
\arguments{
\item{x}{An object of class \code{swaglm}.}

\item{mode}{Character string specifying the mode of the network. Default is "undirected".}
}
\value{
A list of class  \code{swaglm_network} containing:
\itemize{
\item \code{g}: The computed graph object.
\item \code{models}: The selected models matrix.
\item \code{g_simplified_obs}: The simplified network graph.
}
}
\description{
Compute a network representation of the selected models from a \code{swaglm} object
}
\examples{
# Parameters for data generation
set.seed(12345)
n <- 2000
p <- 100

# create design matrix and vector of coefficients
Sigma <- diag(rep(1/p, p))
X <- MASS::mvrnorm(n = n, mu = rep(0, p), Sigma = Sigma)
beta = c(-15,-10,5,10,15, rep(0,p-5))

# --------------------- generate from logistic regression with an intercept of one
z <- 1 + X\%*\%beta
pr <- 1/(1 + exp(-z))
y <- as.factor(rbinom(n, 1, pr))
y = as.numeric(y)-1

# define swag parameters
quantile_alpha = .15
p_max = 20
swag_obj = swaglm::swaglm(X=X, y = y, p_max = p_max, family = stats::binomial(), 
alpha = quantile_alpha, verbose = TRUE, seed = 123)
names(swag_obj)
swag_network = compute_network(swag_obj)
names(swag_network)
}
