% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.swaglm_network.R
\name{plot.swaglm_network}
\alias{plot.swaglm_network}
\title{Plot method for swaglm_network objects}
\usage{
\method{plot}{swaglm_network}(x, bins = 5, scale_edge = NULL, size_range = c(8, 30), ...)
}
\arguments{
\item{x}{An object of class \code{swaglm_network}}

\item{bins}{Number of bins for vertex size discretization (default = 5)}

\item{scale_edge}{Logical; whether to scale the edge width}

\item{size_range}{A numeric vector of length 2 specifying the range of node sizes.}

\item{...}{Additional arguments passed to \code{igraph::plot}}
}
\value{
todo
}
\description{
Visualizes a SWAG network with discretized vertex size, optional edge width scaling,
and edge coloring based on a correlation matrix.
}
