% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.swaglm.R
\name{predict.swaglm}
\alias{predict.swaglm}
\title{print.swaglm}
\usage{
\method{predict}{swaglm}(object, newdata)
}
\arguments{
\item{newdata}{A data.frame that contains the same variable as the training dataset}

\item{x}{An object of class \code{swaglm}.}
}
\value{
TODO.
}
\description{
Predict method for a \code{swaglm} object
}
