% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swaglm}
\alias{swaglm}
\title{sawglm}
\usage{
swaglm(
  X,
  y,
  p_max = 2L,
  family = NULL,
  method = 0L,
  alpha = 0.3,
  verbose = FALSE,
  seed = 123L
)
}
\arguments{
\item{X}{A numeric \code{matrix} of predictors.}

\item{y}{A numeric \code{vector} of responses.}

\item{p_max}{A \code{integer} specifying the maximum dimension to explore}

\item{family}{A \code{family} object. Default is binomial.}

\item{method}{An \code{integer} scalar with value 0 for the column-pivoted QR decomposition, 1 for the unpivoted QR decomposition, 2 for the LLT Cholesky, or 3 for the LDLT Cholesky}

\item{alpha}{a \code{double} specifying the quantile of the criterion used to select models which are employed to construct models to explore at the next dimension}

\item{verbose}{A \code{boolean} used to control verbose}

\item{seed}{A \code{int} that is the random seed used when creating the set of model to explore for the next dimension}
}
\value{
A \code{list} of lists containing:
\itemize{
\item Estimated coefficients per model per dimension.
\item AIC criterion values for models at each dimension.
\item Matrices specifying variable combinations explored at each dimension.
\item Index of selected models.
}
}
\description{
Run the SWAG algorithm supporting Generalized Linear Models specified by a family object and using fastglm library.
}
\examples{
# Parameters for data generation
set.seed(12345)
n <- 2000
p <- 100
# create design matrix and vector of coefficients
Sigma <- diag(rep(1/p, p))
X <- MASS::mvrnorm(n = n, mu = rep(0, p), Sigma = Sigma)
beta = c(-15,-10,5,10,15, rep(0,p-5))

# --------------------- generate from logistic regression with an intercept of one
z <- 1 + X\%*\%beta
pr <- 1/(1 + exp(-z))
y <- as.factor(rbinom(n, 1, pr))
y = as.numeric(y)-1

# define swag parameters
quantile_alpha = .15
p_max = 20
swag_obj = swaglm::swaglm(X=X, y = y, p_max = p_max, family = stats::binomial(), alpha = quantile_alpha, verbose = TRUE, seed = 123)

}
