% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaglm_test.R
\name{swaglm_test}
\alias{swaglm_test}
\title{swaglm_test}
\usage{
swaglm_test(swag_obj, B = 50, verbose = FALSE)
}
\arguments{
\item{swag_obj}{An object of class \code{swaglm}.}

\item{B}{a \code{integer} specifying the number of swag procedures to generate a distribution of the network statistics under the null.}

\item{verbose}{A \code{boolean} used to control verbose}
}
\description{
Compute significance of identified set of variables
}
